backend: only apply diffs to elasticsearch in source crawl

+++ MPCAutofill/cardpicker/sources/update_database.py
@@ -9,6 +9,7 @@ from django.conf import settings
 from django.db import transaction
 
 from cardpicker.constants import DEFAULT_LANGUAGE, MAX_SIZE_MB
+from cardpicker.documents import CardSearch
 from cardpicker.models import Card, CardTypes, Source
 from cardpicker.search.sanitisation import to_searchable
 from cardpicker.sources.api import Folder, Image
@@ -157,34 +158,45 @@ def bulk_sync_objects(source: Source, cards: list[Card]) -> None:
             incoming[identifier].pk = existing[identifier].pk  # this must be explicitly set for bulk_update.
             updated.append(incoming[identifier])
     deleted_ids = existing_ids - incoming_ids
+    deleted = [existing[identifier] for identifier in deleted_ids]
 
     with transaction.atomic():
-        Card.objects.bulk_create(created)
-        Card.objects.bulk_update(
-            updated,
-            # update every field except for `identifier`
-            [
-                "card_type",
-                "name",
-                "priority",
-                "source",
-                "source_verbose",
-                "folder_location",
-                "dpi",
-                "searchq",
-                "extension",
-                "date_created",
-                "date_modified",
-                "size",
-                "tags",
-                "language",
-            ],
-            batch_size=1000,
-        )
-        Card.objects.filter(identifier__in=deleted_ids).delete()
+        if created:
+            Card.objects.bulk_create(created)
+            CardSearch().update(list(created), action="index")
+        if updated:
+            Card.objects.bulk_update(
+                updated,
+                # update every field except for `identifier`
+                [
+                    "card_type",
+                    "name",
+                    "priority",
+                    "source",
+                    "source_verbose",
+                    "folder_location",
+                    "dpi",
+                    "searchq",
+                    "extension",
+                    "date_created",
+                    "date_modified",
+                    "size",
+                    "tags",
+                    "language",
+                ],
+                batch_size=1000,
+            )
+            # as per this thread https://github.com/django-es/django-elasticsearch-dsl/issues/224#issuecomment-551955511
+            # action type "index" is used for indexing new objects as well as updating existing objects
+            CardSearch().update(list(updated), action="index")
+        if deleted_ids:
+            Card.objects.filter(identifier__in=deleted_ids).delete()
+            CardSearch().update(list(deleted), action="delete")
     print(
         f" and done! That took {TEXT_BOLD}{(time.time() - t0):.2f}{TEXT_END} seconds.\n"
-        f"Created {len(created)}, updated {len(updated)}, and deleted {len(deleted_ids)} cards."
+        f"Created {TEXT_BOLD}{len(created)}{TEXT_END}, "
+        f"updated {TEXT_BOLD}{len(updated)}{TEXT_END}, "
+        f"and deleted {TEXT_BOLD}{len(deleted_ids)}{TEXT_END} cards."
     )
